Refactory: move some ApplicationContext functionality.
